requirements.txt
groq==0.4.2
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-groq==0.0.1
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.14
python-dotenv==1.0.1
duckduckgo-search==5.1.0